Simple Test Engaged!
## Changes
- Decouples formatting from Input
- Implements basic test for Formatter
